backend/requirements.txt
# FastAPI 核心框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.11.2
pydantic-settings>=2.0.0
python-multipart==0.0.6

# 异步处理
celery==5.3.4
redis==5.0.1

# 数据库和存储
minio==7.2.0
qdrant-client==1.16.2
sqlalchemy==2.0.23

# 向量和嵌入
numpy>=1.26.0
scipy==1.11.4
scikit-learn==1.3.2
transformers>=4.44.2,<4.50.0  # 兼容 FlagEmbedding (>=4.44.2) 和 CLIP (4.44+)
torch>=2.2.0
FlagEmbedding>=1.2.0  # BGE-M3 稀疏向量编码

# 文档解析
PyMuPDF>=1.24.2
python-docx==1.1.0
markdown==3.5.2
pymupdf4llm==0.0.9
unstructured>=0.18.0

# API 客户端
openai==1.6.1
requests==2.31.0
httpx==0.25.2

# 图像处理
Pillow==10.1.0
opencv-python==4.8.1.78

# 工具库
python-dotenv==1.0.0
aiofiles==23.2.1
cachetools==5.3.2
orjson==3.9.10
jsonschema==4.20.0

# 日志和监控
loguru==0.7.2
prometheus-client==0.19.0

# 数据处理
pandas==2.1.4
tqdm==4.66.1

# 网络和协议
websockets==12.0
sse-starlette==1.8.2

# 安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 开发和测试
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Jupyter (用于调试)
jupyter==1.0.0